Update build to NUKE
Moved the build scripts to NUKE to be more consistent with the Ucommerce
project itself. But also because we have an actually working version of
updating the AssemblyInfo files in NUKE.

To package the payment providers to a zip file just run the `Package`
target which will create a paymentProviders.zip file in the `artifacts/`
directory.

In our pipelines for Ucommerce it is very useful to be able to update
the nuget package for Ucommerce.Core to the latest one built there.
The `UpdateUcommerceCore` target makes this possible. It just needs to
know the directory where the `Ucommerce.Core.*.nupkg` file can be found.

The features of the scripts should be there still in the new build
system. Local deployment is still around as a new target named
`DeployToLocal`

Note: Later when everything is confirmed to work in the Ucommerce
pipeline and the old scripts have been removed from there, we can remove
the old powershell build scripts.

[ch-11660]